requirements.txt
Django==4.0.4
Pillow==9.1.0
PyJWT==2.3.0
loguru==0.6.0
mysqlclient==2.1.0
pytest==7.1.1
requests==2.27.1
rabbitmq-server==3.9.15
redis==6.2.6
celery[redis]==5.2.6
